Make threaded loading safe
Loading resources from a thread other than the main was not safe because a number of resource types schedule a "flush" to get updated according to the loaded values for their properties.

That "flush" generally consists in a deferred call to their `_update()` member function. In other words, calls to that function are added to the `MessageQueue`.

The problem is that if the resource is loaded from another thread, **race conditions happen since the main thread will keep running, flushing the `MessageQueue` and therefore running those calls to `_update()`, while the loading thread may still be dealing with the resource**.

This PR addresses that problem by doing this:
- Adding to `MessageQueue` the ability for the current thread to opt-in to accumulation of deferred calls. When a thread that opted-in decides to opt-out, the accumulated message buffer for it is appended to the regular message queue so they are run when it's safe to do so.
- Making primitive resource loaders (text/binary) use that feature. This way no changes are required on every resource type. To further improve the robustness of this approach, `ResourceInteractiveLoader::poll()` has been refactored so that it's not meant to be overridden (it takes care of the opt-in/opt-out). Derived classes now implement the virtual protected `_poll()` instead, so this new approach is not an added concern to them.